requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.35.76
botocore==1.35.76
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
dacktool==0.0.7
dbstream==0.1.28
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.2.0
django-oauth-toolkit==3.0.1
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
et_xmlfile==2.0.0
google-api-core==2.23.0
google-api-python-client==1.7.11
google-auth==2.36.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==1.2.1
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.66.0
googleauthentication==0.0.18
grpc-google-iam-v1==0.12.7
grpcio==1.68.0
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
jwcrypto==1.5.6
kombu==5.4.2
libcst==1.5.1
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.50
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.4
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==3.0.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
